feat(vscode): session message part refactoring
